Fixes for X11 event handling
Now the events we send to ask for a window refresh are
no longer getting ignored when we use the root window.
Here's the reason, from the XSendEvent(3) manpage:

"If event_mask is the empty set, the event is sent to
the client that created the destination window."

...and that is not us.

Also the ConfigureNotify event is now being used instead
of Expose to detect window resizing.